backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-oauthlib==1.2.4
openai>=1.50.0
requests==2.31.0
python-dateutil==2.8.2
